Transfer of auxiliary functions
